requirements.txt
django==4.2.5
Pillow==10.0.1
environs[django]==9.3.2
python-dateutil==2.8.2